backend/requirements.txt
# Веб-фреймворк
fastapi==0.115.6
uvicorn[standard]==0.34.0

# База данных
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Валидация и настройки
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# Аутентификация
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.20

# AI провайдеры — Anthropic (Claude) + YandexGPT (через httpx)
anthropic==0.42.0

# HTTP клиент (используется YandexProvider)
httpx==0.28.1

# Тестирование
pytest==8.3.4
pytest-asyncio==0.25.0
aiosqlite==0.20.0

# Утилиты
python-dotenv==1.0.1
faker==33.1.0
